Feat/add dry quiet flags (#7)
* Use global logger

* Add DryRun logic

* Add DryRun globally

* Gofmt

+++ filesystem.go
@@ -1,21 +1,20 @@
 package main
 
 import (
-	"log"
 	"fmt"
+	"log"
 )
 
-func MountSingleVolume(drive EbsVol, logger *log.Logger) error {
-	logger.Printf("Mounting single drive: %s", drive.AttachedName)
-	return MountSingleDrive(drive.AttachedName, drive.MountPath, drive.FsType, logger)
+func MountSingleVolume(drive EbsVol) error {
+	log.Printf("Mounting single drive: %s", drive.AttachedName)
+	return MountSingleDrive(drive.AttachedName, drive.MountPath, drive.FsType)
 }
 
-
-func MountSingleDrive(driveName string, mountPath string, desiredFs string, logger *log.Logger) error {
-	if err := checkAndCreateFilesystem(driveName, desiredFs, logger); err != nil {
+func MountSingleDrive(driveName string, mountPath string, desiredFs string) error {
+	if err := checkAndCreateFilesystem(driveName, desiredFs); err != nil {
 		return err
 	}
-	if err := mkdir(mountPath, logger); err != nil {
+	if err := mkdir(mountPath); err != nil {
 		return err
 	}
 	cmd := "mount"
@@ -23,32 +22,32 @@ func MountSingleDrive(driveName string, mountPath string, desiredFs string, logg
 		driveName,
 		mountPath,
 	}
-	logger.Printf("Executing: %s %s", cmd, args)
-	if _, err := ExecuteCommand(cmd, args, logger); err != nil {
-		logger.Printf("%v", err)
+	log.Printf("Executing: %s %s", cmd, args)
+	if _, err := ExecuteCommand(cmd, args); err != nil {
+		log.Printf("%v", err)
 		return err
 	}
 
 	return nil
 }
 
-func mkdir(mountPath string, logger *log.Logger) error {
+func mkdir(mountPath string) error {
 	cmd := "mkdir"
 	args := []string{
 		"-p",
 		mountPath,
 	}
-	logger.Printf("Executing: %s %s", cmd, args)
-	if _, err := ExecuteCommand(cmd, args, logger); err != nil {
-		logger.Printf("%v", err)
+	log.Printf("Executing: %s %s", cmd, args)
+	if _, err := ExecuteCommand(cmd, args); err != nil {
+		log.Printf("%v", err)
 		return err
 	}
 	return nil
 }
 
-func checkAndCreateFilesystem(driveName string, desiredFs string, logger *log.Logger) error {
-	logger.Printf("Checking filesystem on %s", driveName)
-        cmd := "blkid"
+func checkAndCreateFilesystem(driveName string, desiredFs string) error {
+	log.Printf("Checking filesystem on %s", driveName)
+	cmd := "blkid"
 	args := []string{
 		"-o",
 		"value",
@@ -56,27 +55,30 @@ func checkAndCreateFilesystem(driveName string, desiredFs string, logger *log.Lo
 		"TYPE",
 		driveName,
 	}
-	fsOut, err := ExecuteCommand(cmd, args, logger)
+	fsOut, err := ExecuteCommand(cmd, args)
+	if DryRun {
+		return nil
+	}
 	if err != nil {
 		if fsOut.Status == 2 {
-			logger.Printf("Creating fs %s on %s", desiredFs, driveName)
+			log.Printf("Creating fs %s on %s", desiredFs, driveName)
 			cmd = "mkfs"
 			argsCreateFs := []string{
 				"-t",
 				desiredFs,
 				driveName,
 			}
-			if _, err := ExecuteCommand(cmd, argsCreateFs, logger); err != nil {
+			if _, err := ExecuteCommand(cmd, argsCreateFs); err != nil {
 				return err
 			}
 			return nil
 		} else {
-		    logger.Printf("%v", err)
-		    return err
+			log.Printf("%v", err)
+			return err
 		}
 	}
 	switch fsOut.Stdout {
-	case desiredFs+"\n":
+	case desiredFs + "\n":
 		return nil
 	default:
 		return fmt.Errorf("Desired fs: %s, actual fs: %s", desiredFs, fsOut.Stdout)
